feat(benchtool): ensure labelset is used correctly
